eddsa: cleanup and comments
