+++ pkg/eddsa/public_key_test.go
@@ -3,11 +3,25 @@ package eddsa
 import (
 	"crypto/ed25519"
 	"crypto/rand"
+	"crypto/sha512"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/taurusgroup/frost-ed25519/pkg/ristretto"
 )
+
+func newKeyPair(key ed25519.PrivateKey) (*ristretto.Scalar, *PublicKey) {
+	var (
+		sk ristretto.Scalar
+		pk PublicKey
+	)
+	digest := sha512.Sum512(key[:32])
+
+	_, _ = sk.SetBytesWithClamping(digest[:32])
+	pk.pk.ScalarBaseMult(&sk)
+
+	return &sk, &pk
+}
 
 func TestPrivateKey_ToEd25519(t *testing.T) {
 	pkbytes, skBytes, err := ed25519.GenerateKey(rand.Reader)
